backend/requirements.txt
asgiref==3.6.0
astroid==2.15.2
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==3.2
django-import-export==3.0.0b4
django-stubs==1.16.0
django-stubs-ext==0.8.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tablib==3.4.0
tomli==2.0.1
tomlkit==0.11.7
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
